requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
djangorestframework-gis==1.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
packaging==23.1
pip==21.3.1
psycopg2-binary==2.9.6
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
requests==2.31.0
ruamel.yaml.clib==0.2.7
ruamel.yaml==0.17.28
setuptools==60.2.0
sqlparse==0.4.4
typing-extensions==4.6.2
uritemplate==4.1.1
urllib3==2.0.2
wheel==0.37.1
